requirements.txt
scrapy>=2.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
streamlit>=1.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
schedule>=1.2.0
python-dotenv>=1.0.0
